+++ release_notes.md
@@ -17,6 +17,7 @@ this work is available on `master` branch, but has not yet been released.
     * `SocketPair` added (RAII-wrapper for `socketpair()` syscall)
     * `syscallRetry` added wrapper that auto-restarts syscall that failed spuriously
     * `makeNonblocking` wrapper that makes a socket non-blocking
+    * `Epoll` wrapper with functor call on event
 
 
 ## v2.0.0 (2023-08-13)